Kernels for count data

Ruzicka and Bray-Curtis are kernel functions for absolute or relative frequencies and count data. Both kernels have as input a matrix or data.frame with dimension NxD and N>1, D>1, containing strictly non-negative real numbers. Samples should be in the rows. Thus, when working with relative frequencies, `rowSums(X)` should be 1 (or 100, or another arbitrary number) for all rows (samples) of the dataset.

Usage

BrayCurtis(X)

Ruzicka(X)

Arguments

X

Matrix or data.frame that contains absolute or relative frequencies.

Value

Kernel matrix (dimension: NxN).

Details

For more info about these measures, please check Details in ?vegan::vegdist(). Note that, in the vegan help page, "Ruzicka" corresponds to "quantitative Jaccard". `BrayCurtis(X)` gives the same result than `1-vegan::vegdist(X,method="bray")`, and the same with `Ruzicka(data)` and `1-vegan::vegdist(data,method="jaccard")`.

Examples

data <- matrix(rpois(5000,lambda=3),ncol=100,nrow=50)
Kruz <- Ruzicka(data)
Kbray <- BrayCurtis(data)
Kruz[1:5,1:5]
#>           1         2         3         4         5
#> 1 1.0000000 0.5468354 0.5073171 0.5111663 0.4796163
#> 2 0.5468354 1.0000000 0.5207254 0.4974093 0.5181347
#> 3 0.5073171 0.5207254 1.0000000 0.4885496 0.5089059
#> 4 0.5111663 0.4974093 0.4885496 1.0000000 0.4897959
#> 5 0.4796163 0.5181347 0.5089059 0.4897959 1.0000000
Kbray[1:5,1:5]
#>           1         2         3         4         5
#> 1 1.0000000 0.7070376 0.6731392 0.6765189 0.6482982
#> 2 0.7070376 1.0000000 0.6848382 0.6643599 0.6825939
#> 3 0.6731392 0.6848382 1.0000000 0.6564103 0.6745363
#> 4 0.6765189 0.6643599 0.6564103 1.0000000 0.6575342
#> 5 0.6482982 0.6825939 0.6745363 0.6575342 1.0000000