kerntools: Kernel Functions and Tools for Machine Learning Applications

Kernel functions for diverse types of data (including, but not restricted to: nonnegative and real vectors, real matrices, categorical and ordinal variables, sets, strings), plus other utilities like kernel similarity, kernel Principal Components Analysis (PCA) and features' importance for Support Vector Machines (SVMs), which expand other 'R' packages like 'kernlab'.

Author

Maintainer: Elies Ramon eramon@everlyrusher.com (ORCID) [copyright holder]