Linear kernel

`Linear()` computes the inner product between all possible pairs of rows of a matrix or data.frame with dimension NxD.

Usage

Linear(X, cos.norm = FALSE, coeff = NULL)

Arguments

X

Matrix or data.frame that contains real numbers ("integer", "float" or "double").

cos.norm

Should the resulting kernel matrix be cosine normalized? (Defaults: FALSE).

coeff

(optional) A vector of length D that weights each one of the features (columns). When cos.norm=TRUE, `Linear()` first does the weighting and then the cosine-normalization.

Value

Kernel matrix (dimension: NxN).

Examples

dat <- matrix(rnorm(250),ncol=50,nrow=5)
Linear(dat)
#>            [,1]      [,2]      [,3]      [,4]       [,5]
#> [1,] 56.3533820 -1.902548  6.634868  5.378836 -0.6782112
#> [2,] -1.9025476 52.316577  8.875222 -2.050903  5.1256125
#> [3,]  6.6348682  8.875222 57.110595 -8.071251  4.1534043
#> [4,]  5.3788362 -2.050903 -8.071251 36.896757  2.7138702
#> [5,] -0.6782112  5.125612  4.153404  2.713870 39.8589741