Total Sum Scaling

This function transforms a dataset from absolute to relative frequencies (by row or column).

Usage

TSS(X, rows = TRUE)

Arguments

X

Numeric matrix or data.frame of any size containing absolute frequencies.

rows

If TRUE, the operation is done by row; otherwise, it is done by column. (Defaults: TRUE).

Value

A relative frequency matrix or data.frame with the same dimension than X.

Examples

dat <- matrix(rnorm(50),ncol=5,nrow=10)
TSS(dat) #It can be checked that, after scaling, the sum of each row is equal to 1.
#>             [,1]       [,2]       [,3]        [,4]         [,5]
#>  [1,]  0.1792313  1.7862013 -0.4447070  -0.3588527 -0.161872910
#>  [2,]  0.9857786 -0.4544859  0.9839674  -0.2878203 -0.227439758
#>  [3,]  0.5751775  0.7584452 -0.5491987  -0.1692078  0.384783829
#>  [4,]  0.2274961  0.3439082  2.8405000  -1.3568577 -1.055046643
#>  [5,]  0.2667752  0.3638267  0.1160319   0.2544706 -0.001104359
#>  [6,]  0.3659695 -0.5283007  0.6352615   0.5236386  0.003430973
#>  [7,]  0.9700962 -0.2188070  0.3686820   0.9110148 -1.030985981
#>  [8,]  0.1384819  0.3878949  0.3454320  -0.2364140  0.364605206
#>  [9,] 44.7377927 22.3769142 16.4918532 -76.7229752 -5.883584865
#> [10,]  0.4804605 -1.7435773  2.2010283   1.1689823 -1.106893868